fix(testing-1) 오타 정정
