+++ src/main/java/com/cnu/devlog_springboot/DevlogSpringbootApplication.java
@@ -1,0 +1,13 @@
+package com.cnu.devlog_springboot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DevlogSpringbootApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DevlogSpringbootApplication.class, args);
+    }
+
+}
