+++ src/main/java/com/cnu/devlog_springboot/config/WebMVCConfig.java
@@ -1,0 +1,16 @@
+package com.cnu.devlog_springboot.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMVCConfig implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowCredentials(true)
+                .allowedMethods("*")
+                .allowedOriginPatterns("*");
+    }
+}
