+++ src/main/java/com/cnu/devlog_springboot/model/Post.java
@@ -1,0 +1,21 @@
+package com.cnu.devlog_springboot.model;
+
+import com.cnu.devlog_springboot.type.Tag;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Setter
+@Entity(name = "posts")
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED, force = true)
+@Builder
+public class Post {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Integer id;
+    String title;
+    String contents;
+    @Enumerated(value = EnumType.STRING)
+    Tag tag;
+}
