+++ src/main/java/com/cnu/devlog_springboot/type/Tag.java
@@ -1,0 +1,9 @@
+package com.cnu.devlog_springboot.type;
+
+public enum Tag {
+    SPRINGBOOT,
+    JAVA,
+    REACT,
+    OS,
+    NETWORK
+}
