--- src/main/java/com/com/cnu/devlog_springboot/model/request/PostRequest.java
@@ -1,10 +1,0 @@
-package com.com.cnu.devlog_springboot.model.request;
-
-import com.com.cnu.devlog_springboot.type.Tag;
-
-public record PostRequest(
-        String title,
-        String contents,
-        Tag tag
-) {
-}
