--- src/main/java/com/com/cnu/devlog_springboot/repository/PostRepository.java
@@ -1,13 +1,0 @@
-package com.com.cnu.devlog_springboot.repository;
-
-import com.com.cnu.devlog_springboot.model.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
-//    Post save(Post post);
-//    List<Post> findAll();
-//    Optional<Post> findById(Integer postId);
-//    void delete(Post post);
-}
