--- src/main/java/com/com/cnu/devlog_springboot/type/Tag.java
@@ -1,9 +1,0 @@
-package com.com.cnu.devlog_springboot.type;
-
-public enum Tag {
-    SPRINGBOOT,
-    JAVA,
-    REACT,
-    OS,
-    NETWORK
-}
