+++ src/test/java/com/cnu/devlog_springboot/DevlogSpringbootApplicationTests.java
@@ -1,0 +1,13 @@
+package com.cnu.devlog_springboot;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DevlogSpringbootApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
