--- src/test/java/com/com/cnu/devlog_springboot/week1/base/service/PostServiceTest.java
@@ -1,19 +1,0 @@
-package com.com.cnu.devlog_springboot.week1.base.service;
-
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class PostServiceTest {
-
-
-
-    @DisplayName("테스트 코드 실행해보기")
-    @Test
-    void testDoCode() {
-        System.out.println("테스트 코드 실행!");
-    }
-
-}
