--- src/test/java/com/com/cnu/devlog_springboot/week1/base/service/PostServiceWithSpringTest.java
@@ -1,28 +1,0 @@
-package com.com.cnu.devlog_springboot.week1.base.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Optional;
-
-import com.com.cnu.devlog_springboot.model.Post;
-import com.com.cnu.devlog_springboot.service.PostService;
-import com.com.cnu.devlog_springboot.week1.practice.service.fixture.PostFixture;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@ActiveProfiles("dev")
-@SpringBootTest
-public class PostServiceWithSpringTest {
-    @Autowired
-    PostService postService;
-
-    @DisplayName("테스트 코드에서 Spring Component 실행해보기")
-    @Test
-    void testDoCodeRun() {
-        Post post = postService.creatPost(PostFixture.getNormalPostRequest());
-        assertThat(post).isNotNull();
-    }
-}
