--- src/test/java/com/com/cnu/devlog_springboot/week1/practice/service/PostValidServiceTest.java
@@ -1,34 +1,0 @@
-package com.com.cnu.devlog_springboot.week1.practice.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.List;
-
-import com.com.cnu.devlog_springboot.service.valid.PostValidService;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@ActiveProfiles("dev")
-@SpringBootTest
-public class PostValidServiceTest {
-
-    @Autowired
-    PostValidService postValidService;
-
-
-    @DisplayName("post 제목/본문에 비속어가 있나 테스트")
-    @Test
-    void testValidPostIncludeSlang() {
-        // given 시나리오
-        String testTitle = "비속어가 섞인 제목";
-        String testContent = "비속어가 섞인 욕";
-        List<String> slangList = List.of("비속어", "비속어2");
-
-        boolean validPost = postValidService.isSlangInclude(slangList, testTitle, testContent);
-        // then 검증
-        assertThat(validPost).isEqualTo(true);
-    }
-}
